Update subwoofer - fix for Issue #547 on main git repository
Backup is created both on $NAND (directly in MU) and sda0's backup folder. The backup on SD is never used. The new version of this script will attempt to recover the audio datasets from SD card too, if backup on $NAND is unavailable for any reason.

Tested on MU1367-MHI2_ER_VWG13_K4525 - works smooth as butter.